+++ src/full/Agda/Auto/Auto.hs
@@ -8,6 +8,11 @@ module Agda.Auto.Auto
 
 import Prelude hiding (null)
 
+#if __GLASGOW_HASKELL__ <= 708
+import Control.Applicative ( (<$), pure )
+#endif
+
+import Data.Functor
 import Control.Monad.State
 import Data.List hiding (null)
 import qualified Data.Map as Map