+++ src/full/Agda/Syntax/Internal/Pattern.hs
@@ -99,6 +99,7 @@ instance LabelPatVars Pattern DeBruijnPattern Int where
       VarP x       -> do i <- next
                          return $ VarP (DBPatVar x i)
       DotP t       -> DotP t <$ next
+      AbsurdP p    -> AbsurdP <$> labelPatVars p
       ConP c mt ps -> ConP c mt <$> labelPatVars ps
       LitP l       -> return $ LitP l
       ProjP o q    -> return $ ProjP o q
@@ -150,6 +151,7 @@ dbPatPerm' countDots ps = Perm (size ixs) <$> picks
     getIndices (VarP x)      = [Just $ dbPatVarIndex x]
     getIndices (ConP c _ ps) = concatMap (getIndices . namedThing . unArg) ps
     getIndices (DotP _)      = [Nothing | countDots]
+    getIndices (AbsurdP p)   = getIndices p
     getIndices (LitP _)      = []
     getIndices ProjP{}       = []
 
@@ -171,6 +173,7 @@ patternToElim (Arg ai (ConP c cpi ps)) = Apply $ Arg ai $ Con c ci $
       map (argFromElim . patternToElim . fmap namedThing) ps
   where ci = fromConPatternInfo cpi
 patternToElim (Arg ai (DotP t)     ) = Apply $ Arg ai t
+patternToElim (Arg ai (AbsurdP p))   = patternToElim $ Arg ai p
 patternToElim (Arg ai (LitP l)     ) = Apply $ Arg ai $ Lit l
 patternToElim (Arg ai (ProjP o dest)) = Proj o dest
 
@@ -199,6 +202,7 @@ instance MapNamedArg Pattern' where
   mapNamedArg f np =
     case namedArg np of
       VarP  x     -> updateNamedArg VarP $ f $ setNamedArg np x
+      AbsurdP p   -> updateNamedArg AbsurdP $ mapNamedArg f $ setNamedArg np p
       DotP  t     -> setNamedArg np $ DotP t     -- just Haskell type conversion
       LitP  l     -> setNamedArg np $ LitP l     -- ditto
       ProjP o q   -> setNamedArg np $ ProjP o q  -- ditto
@@ -253,6 +257,7 @@ instance PatternLike a (Pattern' a) where
   foldrPattern f p = f p $ case p of
     ConP _ _ ps -> foldrPattern f ps
     VarP _      -> mempty
+    AbsurdP _   -> mempty
     LitP _      -> mempty
     DotP _      -> mempty
     ProjP _ _   -> mempty
@@ -264,6 +269,7 @@ instance PatternLike a (Pattern' a) where
       VarP  _      -> return p
       LitP  _      -> return p
       DotP  _      -> return p
+      AbsurdP _    -> return p
       ProjP _ _    -> return p
 
 -- Boilerplate instances: