+++ src/full/Agda/TypeChecking/Patterns/Match.hs
@@ -181,6 +181,7 @@ matchPattern p u = case (p, u) of
   (VarP x , arg          ) -> return (Yes NoSimplification entry, arg)
     where entry = singleton (dbPatVarIndex x, arg)
   (DotP _ , arg@(Arg _ v)) -> return (Yes NoSimplification empty, arg)
+  (AbsurdP _ , arg@(Arg _ v)) -> return (Yes NoSimplification empty, arg)
   (LitP l , arg@(Arg _ v)) -> do
     w <- reduceB' v
     let arg' = arg $> ignoreBlocking w