+++ src/full/Agda/TypeChecking/Records.hs
@@ -683,6 +683,7 @@ instance NormaliseProjP a => NormaliseProjP (Named_ a) where
 instance NormaliseProjP (Pattern' x) where
   normaliseProjP p@VarP{}        = return p
   normaliseProjP p@DotP{}        = return p
+  normaliseProjP p@AbsurdP{}     = return p
   normaliseProjP (ConP c cpi ps) = ConP c cpi <$> normaliseProjP ps
   normaliseProjP p@LitP{}        = return p
   normaliseProjP (ProjP o d0)    = ProjP o <$> getOriginalProjection d0