+++ src/full/Agda/TypeChecking/Substitute.hs
@@ -220,6 +220,7 @@ instance Apply [NamedArg (Pattern' a)] where
       in  case namedArg p of
             VarP{}  -> recurse
             DotP{}  -> __IMPOSSIBLE__
+            AbsurdP{} -> __IMPOSSIBLE__
             LitP{}  -> __IMPOSSIBLE__
             ConP{}  -> __IMPOSSIBLE__
             ProjP{} -> __IMPOSSIBLE__
@@ -353,6 +354,7 @@ instance Apply Clause where
             VarP (DBPatVar _ i) -> mkSub tm (n - 1) (substP i v' ps) vs `composeS` singletonS i (tm v')
               where v' = raise (n - 1) v
             DotP{}  -> mkSub tm n ps vs
+            AbsurdP p' -> mkSub tm n (setNamedArg p p' : ps) (v : vs)
             ConP c _ ps' -> mkSub tm n (ps' ++ ps) (projections c v ++ vs)
             LitP{}  -> __IMPOSSIBLE__
             ProjP{} -> __IMPOSSIBLE__
@@ -373,6 +375,7 @@ instance Apply Clause where
           case namedArg p of
             VarP (DBPatVar _ i) -> newTel (n - 1) (subTel (size tel - 1 - i) v tel) (substP i (raise (n - 1) v) ps) vs
             DotP{}              -> newTel n tel ps vs
+            AbsurdP{}           -> __IMPOSSIBLE__
             ConP c _ ps'        -> newTel n tel (ps' ++ ps) (projections c v ++ vs)
             LitP{}              -> __IMPOSSIBLE__
             ProjP{}             -> __IMPOSSIBLE__
@@ -731,6 +734,7 @@ instance Subst Term Pattern where
     ConP c mt ps -> ConP c (applySubst rho mt) $ applySubst rho ps
     DotP t       -> DotP $ applySubst rho t
     VarP s       -> p
+    AbsurdP p    -> AbsurdP $ applySubst rho p
     LitP l       -> p
     ProjP{}      -> p
 
@@ -866,6 +870,7 @@ instance Subst DeBruijnPattern DeBruijnPattern where
     VarP x       -> useName (dbPatVarName x) $ lookupS rho $ dbPatVarIndex x
     DotP u       -> DotP $ applyPatSubst rho u
     ConP c ci ps -> ConP c ci $ applySubst rho ps
+    AbsurdP p    -> AbsurdP $ applySubst rho p
     LitP x       -> p
     ProjP{}      -> p
     where